% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drops.R
\name{clear_drops}
\alias{clear_drops}
\title{Clear cache of all "DROP" functions}
\usage{
clear_drops()
}
\description{
The "DROP" functions are "memoised" because the DROP list changes quite slowly.
There is \emph{no need} to update cached data \strong{more than once per hour},
in fact once per day is more than enough in most cases. Automated downloads
must be at least one hour apart. Excessive downloads may result in your IP
being firewalled from the Spamhaus website.\cr\cr
If you are using them from
a long-running R session and need to fetch the list, use the this function
to clear the cache of all "DROP" get functions.
}
