% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drops.R
\name{get_asndrop}
\alias{get_asndrop}
\title{Spamhaus ASN DROP List (ASN-DROP)}
\usage{
get_asndrop()
}
\description{
Retrieves the Spamhaus ASN DROP List (ASN-DROP)
}
\details{
ASN-DROP contains a list of Autonomous System Numbers controlled by spammers
or cyber criminals, as well as "hijacked" ASNs. ASN-DROP can be used to filter
BGP routes which are being used for malicious purposes.
}
\note{
This function is "memoised" because the DROP list changes quite slowly.
There is \emph{no need} to update cached data \strong{more than once per hour},
in fact once per day is more than enough in most cases. Automated downloads
must be at least one hour apart. Excessive downloads may result in your IP
being firewalled from the Spamhaus website.\cr\cr
If you are using this from
a long-running R session and need to fetch the list, use the \code{\link[=clear_drops]{clear_drops()}}
function to clear the cache of all "DROP" get functions.
}
\references{
\url{https://www.spamhaus.org/drop/}; \url{https://www.spamhaus.org/faq/section/}
}
