% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drops.R
\name{get_dropv6}
\alias{get_dropv6}
\title{Spamhaus IPv6 DROP List (DROPv6)}
\usage{
get_dropv6()
}
\description{
Retrieves the Spamhaus IPv6 DROP List (DROPv6)
}
\details{
The DROPv6 list includes IPv6 ranges allocated to spammers or cyber criminals.
DROPv6 will only include IPv6 netblocks allocated directly by an established
Regional Internet Registry (RIR) or National Internet Registry (NIR) such as
ARIN, RIPE, AFRINIC, APNIC, LACNIC or KRNIC or direct RIR allocations.
}
\note{
This function is "memoised" because the DROP list changes quite slowly.
There is \emph{no need} to update cached data \strong{more than once per hour},
in fact once per day is more than enough in most cases. Automated downloads
must be at least one hour apart. Excessive downloads may result in your IP
being firewalled from the Spamhaus website.\cr\cr
If you are using this from
a long-running R session and need to fetch the list, use the \code{\link[=clear_drops]{clear_drops()}}
function to clear the cache of all "DROP" get functions.
}
\references{
\url{https://www.spamhaus.org/drop/}; \url{https://www.spamhaus.org/faq/section/}
}
