% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hormel-package.R
\docType{package}
\name{hormel}
\alias{hormel}
\alias{hormel-package}
\title{Retrieve and Process 'Spamhaus' Zone/Host Metadata}
\description{
The 'Spamhaus Project' is an international nonprofit organization that
racks spam and related cyber threats such as phishing, malware and botnets,
provides realtime actionable and highly accurate threat intelligence to the
Internet's major networks, corporations and security vendors, and works with
law enforcement agencies to identify and pursue spam and malware sources
worldwide. Methods are provided to retreive and process 'Spamhaus' data.
}
\author{
Bob Rudis (bob@rud.is)
}
